feat: using ai for Voice to DB command (#11)

+++ src/tfm/model.py
@@ -1,12 +1,13 @@
 from dataclasses import dataclass, field
 from datetime import datetime
+from decimal import Decimal
 
 
 @dataclass
 class User:
     id: int
-    initial_balance: float
-    current_balance: float
+    initial_balance: Decimal
+    current_balance: Decimal
     created_at: datetime
     updated_at: datetime
     transactions_types: list[str] = field(default_factory=list)
@@ -15,7 +16,7 @@ class User:
 @dataclass
 class Transaction:
     user_id: int
-    amount: float
+    amount: Decimal
     transaction_date: datetime
     transaction_type: str
     description: str
